+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM node:8.9-stretch
+FROM node:10-stretch
 
 RUN apt-get update &&  \
     apt-get install -y git wget curl unzip ruby build-essential && \
